reorganized administration frontend code
