+++ contentcuration/contentcuration/frontend/administration/components/ConfirmationDialog.vue
