+++ contentcuration/contentcuration/frontend/administration/pages/AdministrationIndex.vue
