+++ contentcuration/contentcuration/frontend/administration/pages/Channels/ChannelTableHeader.vue
@@ -3,10 +3,10 @@
   <v-container fluid grid-list-xl>
     <v-layout wrap align-center justify-end>
       <v-flex shrink>
-        <h1>{{ totalItems }} Users</h1>
+        <h1>{{ totalItems }} Channels</h1>
       </v-flex>
       <v-flex grow>
-        <UserActionsBulk :selected="selected" />
+        <ChannelActionsBulk :selected="selected" />
       </v-flex>
       <v-flex md3 d-flex>
         <v-select
@@ -14,16 +14,17 @@
           :items="filterTypes"
           item-text="label"
           item-value="key"
-          label="User Type"
+          label="Channel Type"
           @change="filter"
         />
       </v-flex>
       <v-flex xs12 md4>
         <v-text-field
           v-model="pagination.search"
-          label="Search for a user..."
+          label="Search for a channel..."
           prepend-inner-icon="search"
           clearable
+
           @change="search"
           @click:clear="clearSearch"
         />
@@ -37,13 +38,13 @@
 
 <script>
 
-  import { filterMixin } from '../mixins';
-  import { userFilterTypes } from '../constants';
-  import UserActionsBulk from './UserActionsBulk';
+  import { filterMixin } from '../../mixins';
+  import { channelFilterTypes } from '../../constants';
+  import ChannelActionsBulk from './ChannelActionsBulk';
 
   export default {
-    name: 'UserTableHeader',
-    components: { UserActionsBulk },
+    name: 'ChannelTableHeader',
+    components: { ChannelActionsBulk },
     mixins: [filterMixin],
     props: {
       pagination: Object,
@@ -51,7 +52,19 @@
       totalItems: Number,
     },
     data: () => ({
-      filterTypes: userFilterTypes,
+      searchQuery: '',
+      filterTypes: channelFilterTypes,
+      sortBy: [
+        { key: 'joined', label: 'Date Joined' },
+        { key: 'email', label: 'Email' },
+        { key: 'first_name', label: 'First Name' },
+        { key: 'last_name', label: 'Last Name' },
+        { key: 'editing', label: 'Channels Editing' },
+      ],
+      order: [
+        { key: 'descending', label: 'Descending' },
+        { key: 'ascending', label: 'Ascending' },
+      ],
     }),
   };
