+++ contentcuration/contentcuration/frontend/administration/pages/Users/UserInfo.vue
